chat/requirements.txt
fastapi
uvicorn
websockets
pyjwt